fixing tons of lints

+++ videocall-client/src/encode/microphone_encoder.rs
@@ -92,7 +92,7 @@ impl MicrophoneEncoder {
                     if enabled.load(Ordering::Acquire) {
                         // Only update if change is greater than threshold
                         let current = current_bitrate.load(Ordering::Relaxed) as f64;
-                        let new = bitrate as f64;
+                        let new = bitrate;
                         let percent_change = (new - current).abs() / current;
 
                         if percent_change > BITRATE_CHANGE_THRESHOLD {
